Updated the mdms call

+++ municipal-services/pt-calculator-v2/src/main/java/org/egov/pt/calculator/web/models/property/PropertyDetail.java
@@ -95,10 +95,10 @@ public class PropertyDetail   {
         private Long noOfFloors;
 
         @JsonProperty("landArea")
-        private Double landArea;
+        private BigDecimal landArea;
 
         @JsonProperty("buildUpArea")
-        private Double buildUpArea;
+        private BigDecimal buildUpArea;
 
         @JsonProperty("units")
         @Valid
